[GTSAM catkin] Use build directory for building the package
Currently, building GTSAM catkin downloads into the source directory a
bunch of repos which later cause issues when re-building the package.

To ensure that a catkin clean removes any residual, I have moved the
files that are being downloaded into the build folder for this package.

Tested on local machine